agent_finder/requirements.txt
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openpyxl>=3.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
ddgs>=9.0.0
dnspython>=2.6.0
